+++ app/frontend/src/components/molecules/snapshot/DetailDescription.tsx
