+++ app/frontend/src/components/molecules/snapshot/DetailTitle.tsx
