+++ app/frontend/src/components/molecules/snapshot/SnapshotViewContent.tsx
@@ -1,19 +1,21 @@
-import { useEffect, useMemo, useRef } from "react";
-import { privateScoreItemUrlGen } from "../../global";
+import React, { useEffect, useMemo, useRef } from "react";
+import { privateScoreItemUrlGen } from "../../../global";
 import Viewer from "viewerjs";
-import { ScorePage } from "../../ScoreClientV2";
+import { ScorePage } from "../../../ScoreClientV2";
 import { useHistory } from "react-router-dom";
 
-export interface ViewContentProps {
+export interface SnapshotViewContentProps {
   ownerId?: string;
   scoreId?: string;
+  snapshotId?: string;
   pages?: ScorePage[];
   pageIndex?: number;
 }
 
-export function ViewContent(props: ViewContentProps) {
+export function SnapshotViewContent(props: SnapshotViewContentProps) {
   const _ownerId = props.ownerId;
   const _scoreId = props.scoreId;
+  const _snapshotId = props.snapshotId;
   const _pages = props.pages ?? [];
   const _pageIndex = props.pageIndex;
   const ulRef = useRef<HTMLUListElement>(null);
@@ -26,7 +28,6 @@ export function ViewContent(props: ViewContentProps) {
   useEffect(() => {
     if (!ulRef.current) return;
     if (!viewerContainerRef.current) return;
-    console.log("init view");
     const viewer = new Viewer(ulRef.current, {
       url: "data-original",
       loop: false,
@@ -54,9 +55,12 @@ export function ViewContent(props: ViewContentProps) {
         if (index < 0 || _pages.length <= index) return;
         const page = _pages[index];
         // 画像を変更したときに URL も変更する
-        history.push(`/scores/${_scoreId}/page/${page.id}`);
+        history.push(
+          `/scores/${_scoreId}/snapshot/${_snapshotId}/page/${page.id}`
+        );
       },
     });
+    console.log(viewerContainerRef.current);
     viewerRef.current = viewer;
 
     return () => {
