+++ app/frontend/src/components/organisms/ScoreDetailContent.tsx
@@ -1,8 +1,8 @@
 import { Button, createStyles, makeStyles, Theme } from "@material-ui/core";
 import { ArrowBack } from "@material-ui/icons";
 import { useHistory, useParams } from "react-router";
-import LatestScoreContent from "../atoms/LatestScoreContent";
-import ScoreSnapshotContent from "../atoms/ScoreSnapshotContent";
+import LatestScoreContent from "../molecules/detail/LatestScoreContent";
+import ScoreSnapshotContent from "../molecules/snapshot/ScoreSnapshotContent";
 import SnapshotNameList from "../atoms/SnapshotNameList";
 
 const useStyles = makeStyles((theme: Theme) =>
